create & separate out MoreOptionsTabs component
